SWITCHYARD-1308 Implement binding.sca and extensions
