+++ pom.xml
@@ -46,7 +46,7 @@
     <modules>
         <module>common</module>
         <module>test</module>
-        <module>remote</module>
+        <module>sca</module>
         <module>bean</module>
         <module>rules</module>
         <module>bpm</module>
