--- remote/src/main/java/org/switchyard/component/remote/config/model/RemoteBindingModel.java
@@ -1,37 +1,0 @@
-/* 
- * JBoss, Home of Professional Open Source 
- * Copyright 2012 Red Hat Inc. and/or its affiliates and other contributors
- * as indicated by the @author tags. All rights reserved. 
- * See the copyright.txt in the distribution for a 
- * full listing of individual contributors.
- *
- * This copyrighted material is made available to anyone wishing to use, 
- * modify, copy, or redistribute it subject to the terms and conditions 
- * of the GNU Lesser General Public License, v. 2.1. 
- * This program is distributed in the hope that it will be useful, but WITHOUT A 
- * WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A 
- * PARTICULAR PURPOSE.  See the GNU Lesser General Public License for more details. 
- * You should have received a copy of the GNU Lesser General Public License, 
- * v.2.1 along with this distribution; if not, write to the Free Software 
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, 
- * MA  02110-1301, USA.
- */
-package org.switchyard.component.remote.config.model;
-
-import org.switchyard.config.model.composite.BindingModel;
-
-/**
- * Representation of a 'binding.remote' element.
- */
-public interface RemoteBindingModel extends BindingModel {
-    
-    /**
-     * The "remote" namespace.
-     */
-    public static final String DEFAULT_NAMESPACE = "urn:switchyard-component-remote:config:1.0";
-    
-    /**
-     * The 'remote' binding type.
-     */
-    String REMOTE = "remote";
-}
