--- remote/src/main/resources/META-INF/services/org.switchyard.deploy.Component
@@ -1,1 +1,0 @@
-org.switchyard.component.remote.deploy.RemoteComponent
