--- remote/src/main/resources/org/switchyard/component/remote/config/model/v1/remote-v1.xsd
@@ -1,37 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-JBoss, Home of Professional Open Source
-Copyright 2011 Red Hat Inc. and/or its affiliates and other contributors
-as indicated by the @authors tag. All rights reserved.
-See the copyright.txt in the distribution for a
-full listing of individual contributors.
-
-This copyrighted material is made available to anyone wishing to use,
-modify, copy, or redistribute it subject to the terms and conditions
-of the GNU Lesser General Public License, v. 2.1.
-This program is distributed in the hope that it will be useful, but WITHOUT A
-WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A
-PARTICULAR PURPOSE.  See the GNU Lesser General Public License for more details.
-You should have received a copy of the GNU Lesser General Public License,
-v.2.1 along with this distribution; if not, write to the Free Software
-Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston,
-MA  02110-1301, USA.
--->
-<schema xmlns="http://www.w3.org/2001/XMLSchema"
-        targetNamespace="urn:switchyard-component-remote:config:1.0"
-        xmlns:remote="urn:switchyard-component-remote:config:1.0"
-        xmlns:sca="http://docs.oasis-open.org/ns/opencsa/sca/200912"
-        elementFormDefault="qualified">
-
-    <import namespace="http://docs.oasis-open.org/ns/opencsa/sca/200912"/>
-    
-    <!--  clojure:implementation.clojure -->
-    <element name="binding.remote" type="remote:RemoteBindingType" substitutionGroup="sca:binding"/>
-    
-    <complexType name="RemoteBindingType">
-        <complexContent>
-            <extension base="sca:Binding"/>
-        </complexContent>
-    </complexType>
-    
-</schema>
