+++ sca/src/main/resources/META-INF/services/org.switchyard.deploy.Component
@@ -1,0 +1,1 @@
+org.switchyard.component.sca.deploy.SCAComponent
