+++ tools/forge/sca/src/main/resources/META-INF/beans.xml
